[refactoring] moving things around
* create new cpp geometry, rendering and
objReader
* renamed and split tests
* added unique way to include opengl
