+++ src/ObjModel.cpp
@@ -1,28 +1,15 @@
 /**
- * @file ObjModel.cpp
- * @author  Simone Gasparini <simone.gasparini@enseeiht.fr>
- * @version 1.0
- *
- * @section LICENSE
- *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- *
- * @section DESCRIPTION
- * 
- * Simple Class to load and draw 3D objects from OBJ files
- * Using triangles and normals as static object. No texture mapping. 
- * OBJ files must be triangulated!!!
- *
  */
 
 #include "ObjModel.hpp"
+#include "objReader.hpp"
+#include "geometry.hpp"
 
 #include <iostream>
 #include <fstream>
-#include <cstdio>
-#include <cstring>
 #include <cassert>
 
 #include <sstream>
@@ -32,302 +19,7 @@
 
 using namespace std;
 
-/**
- * Calculate the normal of a triangular face defined by three points
- *
- * @param[in] v1 the first vertex of the face
- * @param[in] v2 the second vertex of the face
- * @param[in] v3 the third vertex of the face
- * @return the normal of the face as a normalized vector
- * @note the normal is normalized
- */
-vec3d computeNormal( const point3d& v1, const point3d& v2, const point3d& v3)
-{
-    //**************************************************
-    // compute the cross product between two edges of the triangular face
-    //**************************************************
-    //++ vec3d norm;
-    auto norm = (v1 - v2).cross( v1 - v3 );  //!!
 
-    //**************************************************
-    // remember to normalize the result before returning it
-    //**************************************************
-    norm.normalize( );  //!!
-
-    return norm;
-}
-
-/**
- * Load the OBJ data from file
- * @param[in] filename The name of the OBJ file to load
- * @param[out] vertices The list of vertices
- * @param[out] mesh The list of faces
- * @param[out] normals The list of normals
- * @param[out] bb The bounding box of the object
- * @return true if everything went well, false otherwise
- */
-bool load(const std::string& filename, std::vector<point3d>& vertices, std::vector<face>& mesh, std::vector<vec3d>& normals, BoundingBox& bb)
-{
-    string line;
-    ifstream objFile( filename );
-
-    // If obj file is not open return (e.g. file does not exist
-    if (! objFile.is_open( ) )
-    {
-        std::cerr << "Unable to open file " << filename << std::endl;
-        return false;
-    }
-
-    // Start reading file data
-    while( !objFile.eof( ) )
-    {
-        // Get a line from file
-        getline( objFile, line );
-
-        // If the first character is a simple 'v'...
-//            PRINTVAR( line );
-        if ( (line[0] == 'v') && (line[1] == ' ') ) // to drop all the vn and vn lines
-        {
-//                PRINTVAR( line );
-            // Read 3 floats from the line:  X Y Z and store them in the corresponding place in _vertices
-            const point3d p = parseVertexString(line);
-
-            //**************************************************
-            // add the new point to the list of the vertices
-            // and its normal to the list of normals: for the time
-            // being it is a [0, 0 ,0] normal.
-            //**************************************************
-            vertices.push_back( p );  //!!
-            normals.emplace_back();  //!!
-
-            // update the bounding box, if it is the first vertex simply
-            // set the bb to it
-            if (vertices.size( ) == 1 )
-            {
-                bb.set( p );
-            }
-            else
-            {
-                // otherwise add the point
-                bb.add( p );
-            }
-
-        }
-        // If the first character is a 'f'...
-        if ( line[0] == 'f' )
-        {
-            face t = parseFaceString( line);
-
-            //**************************************************
-            // correct the indices: OBJ starts counting from 1, in C the arrays starts at 0...
-            //**************************************************
-            t -= 1;  //!!
-
-            //**************************************************
-            // add it to the mesh
-            //**************************************************
-            mesh.push_back( t );  //!!
-
-            //*********************************************************************
-            //  Compute the normal of the face
-            //*********************************************************************
-            const vec3d norm = computeNormal(vertices[ t.v1], vertices[t.v2], vertices[t.v3]);  //!!
-
-            //*********************************************************************
-            // Sum the normal of the face to each vertex normal
-            //*********************************************************************
-
-            normals[t.v1] += (vec3d( norm ) * angleAtVertex(vertices[ t.v1], vertices[t.v2], vertices[t.v3] ));  //!!
-            normals[t.v2] += (vec3d( norm ) * angleAtVertex(vertices[ t.v2], vertices[t.v1], vertices[t.v3] ));  //!!
-            normals[t.v3] += (vec3d( norm ) * angleAtVertex(vertices[ t.v3], vertices[t.v1], vertices[t.v2] ));  //!!
-
-        }
-    }
-
-    cerr << "Found :\n\tNumber of triangles (_indices) " << mesh.size( ) << "\n\tNumber of Vertices: " << vertices.size( ) << "\n\tNumber of Normals: " << normals.size( ) << endl;
-//        PRINTVAR( mesh );
-//        PRINTVAR( vertices );
-//        PRINTVAR( normals );
-
-    //*********************************************************************
-    // normalize the normals of each vertex
-    //*********************************************************************
-    for(auto &normal : normals)  //<!!
-    {
-        normal.normalize();
-    }  //>!!
-
-//        PRINTVAR( normals );
-
-    // Close OBJ file
-    objFile.close();
-
-
-    cout << "Object loaded with " << vertices.size( ) << " vertices and " << mesh.size( ) << " faces" << endl;
-    cout << "Bounding box : pmax=" << bb.pmax << "  pmin=" << bb.pmin << endl;
-    return true;
-}
-
-bool ObjModel::load(const std::string& filename)
-{
-    return ::load(filename, _vertices, _mesh, _normals, _bb);
-}
-
-/**
- * Draw the wireframe of the model
- * 
- * @param vertices The list of vertices
- * @param mesh The mesh as a list of faces, each face is a tripleIndex of vertex indices 
- * @param params The rendering parameters
- */
-void ObjModel::drawWireframe( const std::vector<point3d> &vertices, const std::vector<face> &mesh, const RenderingParameters &params ) const
-{
-    //**************************************************
-    // we first need to disable the lighting in order to
-    // draw colored segments
-    //**************************************************
-    glDisable( GL_LIGHTING );
-
-    // if we are displaying the object with colored faces
-    if ( params.solid )
-    {
-        // use black ticker lines
-        glColor3f( 0, 0, 0 );
-        glLineWidth( 2 );
-    }
-    else
-    {
-        // otherwise use white thinner lines for wireframe only
-        glColor3f( .8f, .8f, .8f );
-        glLineWidth( .21f );
-    }
-
-    //**************************************************
-    // for each face of the mesh...
-    //**************************************************
-    for(const auto &face : mesh)  //!!
-    {
-        //**************************************************
-        // draw the contour of the face as a  GL_LINE_LOOP
-        //**************************************************
-        glBegin( GL_LINE_LOOP );  //<!!
-            glVertex3fv( (float*) &vertices[face.v1] );
-
-            glVertex3fv( (float*) &vertices[face.v2] );
-
-            glVertex3fv( (float*) &vertices[face.v3] );
-        glEnd( );  //>!!
-    }
-
-    //**************************************************
-    // re-enable the lighting
-    //**************************************************
-    glEnable( GL_LIGHTING );
-}
-
-
-/**
- * Draw the faces using the computed normal of each face
- * 
- * @param[in] vertices The list of vertices
- * @param[in] mesh The list of face, each face containing the indices of the vertices
- * @param[in] params The rendering parameters
- */
-void ObjModel::drawFlatFaces( const std::vector<point3d> &vertices, const std::vector<face> &mesh, const RenderingParameters &params ) const
-{
-    // shading model to use
-    if ( params.smooth )
-    {
-        glShadeModel( GL_SMOOTH );
-    }
-    else
-    {
-        glShadeModel( GL_FLAT );
-    }
-
-    //**************************************************
-    // for each face
-    //**************************************************
-    for(const auto &face : mesh)  //!!
-    {
-        //**************************************************
-        // Compute the normal to the face and then draw the
-        // faces as GL_TRIANGLES assigning the proper normal
-        //**************************************************
-        glBegin( GL_TRIANGLES );  //<!!
-
-            const vec3d n = computeNormal( vertices[face.v1], vertices[face.v2], vertices[face.v3]);
-            glNormal3fv( (float*) &n );
-
-            glVertex3fv( (float*) &vertices[face.v1] );
-
-            glVertex3fv( (float*) &vertices[face.v2] );
-
-            glVertex3fv( (float*) &vertices[face.v3] );
-
-        glEnd( );  //>!!
-    }
-}
-
-
-
-/**
- * Draw the model using the vertex indices
- * 
- * @param vertices The vertices 
- * @param indices The list of the faces, each face containing the 3 indices of the vertices
- * @param vertexNormals The list of normals associated to each vertex
- * @param params The rendering parameters
- */
-void ObjModel::drawSmoothFaces( const std::vector<point3d> &vertices, 
-                                const std::vector<face> &mesh,
-                                std::vector<vec3d> &vertexNormals,
-                                const RenderingParameters &params ) const
-{
-    if ( params.smooth )
-    {
-        glShadeModel( GL_SMOOTH );
-    }
-    else
-    {
-        glShadeModel( GL_FLAT );
-    }
-    //****************************************
-    // Enable vertex arrays
-    //****************************************
-    glEnableClientState( GL_VERTEX_ARRAY );  //!!
-
-    //****************************************
-    // Enable normal arrays
-    //****************************************
-    glEnableClientState( GL_NORMAL_ARRAY );  //!!
-
-    //****************************************
-    // Normal pointer to normal array
-    //****************************************
-    glNormalPointer( GL_FLOAT, 0, (float*) &vertexNormals[0] );  //!!
-
-    //****************************************
-    // Vertex pointer to Vertex array
-    //****************************************
-    glVertexPointer( COORD_PER_VERTEX, GL_FLOAT, 0, (float*) &vertices[0] );  //!!
-
-    //****************************************
-    // Draw the faces
-    //****************************************
-    glDrawElements( GL_TRIANGLES, static_cast<GLsizei>( mesh.size( )) * VERTICES_PER_TRIANGLE, GL_UNSIGNED_INT, (idxtype*) & mesh[0] );  //!!
-
-    //****************************************
-    // Disable vertex arrays
-    //****************************************
-    glDisableClientState( GL_VERTEX_ARRAY );  //!!
-
-    //****************************************
-    // Disable normal arrays
-    //****************************************
-    glDisableClientState( GL_NORMAL_ARRAY );  //!!
-
-}
 
 /**
  * Compute the subdivision of the input mesh by applying one step of the Loop algorithm
