+++ src/openglAll.hpp
@@ -1,0 +1,17 @@
+#pragma once
+
+// for mac osx
+#ifdef __APPLE__
+#include <OpenGL/gl.h>
+#include <OpenGL/glu.h>
+#include <GLUT/glut.h>
+#else
+// only for windows
+#ifdef _WIN32
+#include <windows.h>
+#endif
+// for windows and linux
+#include <GL/gl.h>
+#include <GL/glu.h>
+#include <GL/freeglut.h>
+#endif
