+++ src/tests/test_objReader.cpp
@@ -7,8 +7,9 @@
 #define BOOST_TEST_DYN_LINK
 
 #include <boost/test/unit_test.hpp>
+#include <objReader.hpp>
 #include <core.hpp>
-#include <ObjModel.hpp>
+
 
 #include <map>
 #include <string>
@@ -67,30 +68,6 @@ BOOST_AUTO_TEST_CASE(test_parse_vertex)
             BOOST_CHECK_CLOSE(vertex.z, line.second.value().z, 0.0001);
         }
     }
-}
-
-BOOST_AUTO_TEST_CASE(test_compute_normal)
-{
-    const auto normal = computeNormal({0, 0, 0}, {1, 0, 0}, {0, 1, 0});
-    BOOST_CHECK_CLOSE(normal.x, .0f, 0.0001f);
-    BOOST_CHECK_CLOSE(normal.y, .0f, 0.0001f);
-    BOOST_CHECK_CLOSE(normal.z, 1.f, 0.0001f);
-    BOOST_CHECK_CLOSE(normal.norm(), 1.f, 0.0001f);
-}
-
-BOOST_AUTO_TEST_CASE(test_angleAtVertex)
-{
-    const auto c45 = static_cast<float>(std::cos(M_PI_4));
-    const auto s45 = static_cast<float>(std::sin(M_PI_4));
-    BOOST_CHECK_CLOSE(angleAtVertex({0, 0, 0}, {1, 0, 0}, {0, 1, 0}), M_PI_2, 0.0001f);
-    BOOST_CHECK_CLOSE(angleAtVertex({0, 0, 0}, {1, 0, 0}, {c45, s45, 0}), M_PI_4, 0.0001f);
-    BOOST_CHECK_CLOSE(angleAtVertex({0, 0, 0}, {1, 0, 0}, {-c45, s45, 0}), 3.f*M_PI_4, 0.0001f);
-    BOOST_CHECK_CLOSE(angleAtVertex({0, 0, 0}, {1, 0, 0}, {-c45, -s45, 0}), 3.f*M_PI_4, 0.0001f);
-    BOOST_CHECK_CLOSE(angleAtVertex({0, 0, 0}, {c45, s45, 0}, {1, 0, 0}), M_PI_4, 0.0001f);
-    BOOST_CHECK_CLOSE(angleAtVertex({0, 0, 0}, {c45, -s45, 0}, {1, 0, 0}), M_PI_4, 0.0001f);
-    BOOST_CHECK_CLOSE(angleAtVertex({0, 0, 0}, {-c45, -s45, 0}, {1, 0, 0}), 3.f*M_PI_4, 0.0001f);
-    BOOST_CHECK_CLOSE(angleAtVertex({0, 0, 0}, {1, 0, 0}, {1, 0, 0}), .0f, 0.0001f);
-
 }
